requirements.txt
apache-airflow==2.7.1
psycopg2-binary
requests
beautifulsoup4
python-dotenv
pyspark==3.3.0
apache-airflow-providers-apache-spark